+++ packages/core/src/core/form/LmButton.tsx
@@ -11,11 +11,12 @@ export const LmButton = forwardRef(function LmButtonFunc(
   { loading, colorVariant, ...props }: LmButtonProps,
   ref
 ) {
+  const { theme } = props
   return (
     <Button
       {...props}
       ref={ref as TamaguiComponent}
-      theme={colorVariant ? colormap[colorVariant] : props.theme}
+      theme={colorVariant ? colormap[colorVariant] : theme}
       {...(loading && {
         icon: <Spinner />,
         disabled: true,