+++ packages/core/src/core/themeMappings.ts
@@ -1,11 +1,11 @@
-import {ThemeProps} from "tamagui";
+import { ThemeName } from 'tamagui'
 
 export type ThemeColors = 'primary' | 'secondary' | 'info' | 'warning' | 'success' | 'error'
-export const colormap: { [k in ThemeColors]: ThemeProps['name'] } = {
-    success: 'green_alt1',
-    error: 'red_alt1',
-    info: 'blue_alt1',
-    primary: 'blue_alt1',
-    secondary: 'gray_alt2',
-    warning: 'orange_alt2'
+export const colormap: { [k in ThemeColors]: ThemeName } = {
+  success: 'green',
+  error: 'red',
+  info: 'blue',
+  primary: 'blue',
+  secondary: 'gray' as ThemeName,
+  warning: 'orange',
 }