+++ packages/form/types/LmSwitch.d.ts
@@ -1,7 +1,8 @@
-import { SwitchProps } from "tamagui";
+import { SwitchProps, SwitchThumbProps } from 'tamagui';
 export type LmSwitchProps = SwitchProps & {
     labelLeft?: string;
     labelRight?: string;
+    thumbProps?: SwitchThumbProps;
 };
-export declare function LmSwitch({ labelLeft, labelRight, space, size, ...rest }: LmSwitchProps): import("react/jsx-runtime").JSX.Element;
+export declare function LmSwitch({ labelLeft, labelRight, space, size, thumbProps, ...rest }: LmSwitchProps): import("react/jsx-runtime").JSX.Element;
 //# sourceMappingURL=LmSwitch.d.ts.map