+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.0.1 (2023-04-02)
 
+## 1.3.0
+
+### Minor Changes
+
+- Fix Button colors and upgrade to latest Tamagui version including NextJS and Storybook
+
 ## 1.2.0
 
 ### Minor Changes